requirements.txt
# Web Framework
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.1
gunicorn==20.1.0
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0

# Supabase Integration
supabase==0.7.1
gotrue==0.5.0
postgrest==0.10.2
realtime==0.0.8

# Database Drivers
psycopg2-binary==2.9.3
pymongo==4.1.1
neo4j==4.4.5
mysql-connector-python==8.0.28
sqlalchemy==1.4.39

# Docker Integration
docker==5.0.3

# Utilities
python-dotenv==0.19.1
requests==2.26.0
python-dateutil==2.8.2
retrying==1.3.3
tenacity==8.0.1

# Testing
pytest==6.2.5
pytest-cov==2.12.1
pytest-mock==3.7.0

# Logging and Monitoring
structlog==21.5.0
pyformance==0.4
